pyproject.toml
[project]
name = "rebook2"
version = "0.1.0"
description = "Python project scaffolded for uv, Ruff, and ty."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "easyocr>=1.7.2",
    "glmocr",
    "matplotlib>=3.10.8",
    "mplcursors>=0.7",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.1",
    "pypdfium2>=5.6.0",
    "python-doctr>=1.0.1",
    "torch>=2.8.0",
    "transformers>=5.0.0",
    "vllm[none]>=0.11.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]

[tool.uv]
package = false

[tool.uv.sources]
glmocr = { git = "https://github.com/zai-org/GLM-OCR" }

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ty]
